# Rewrite of the decoder