… generated_for_cmake/google/protobuf/descriptor.upb.h
… tests/conformance_upb.c
… tests/test_generated_code.c
… upb/encode.c
… upb/msg.c
… upb/msg.h
… upb/port_def.inc
… upb/port_undef.inc
… upb/reflection.c
… upb/upb.c
… upb/upb.h
… upbc/generator.cc